+++ src/errors/mod.rs
@@ -1,6 +1,7 @@
 mod capacity_mismatch;
 mod collection_collision;
 mod collection_error;
+mod item_collision;
 mod result_collection_error;
 mod tuple_collection_error;
 mod tuple_extension_error;
@@ -9,6 +10,7 @@ mod unzip_error;
 pub use capacity_mismatch::*;
 pub use collection_collision::*;
 pub use collection_error::*;
+pub use item_collision::*;
 pub use result_collection_error::*;
 pub use tuple_collection_error::*;
 pub use tuple_extension_error::*;