+++ src/impls/maps/hashbrown.rs
@@ -87,3 +87,18 @@ where
         .map_err(|(staging_map, kvp)| CollectionCollision::new(iter, staging_map, kvp))
     }
 }
+
+impl<K: Eq + Hash, V, S: BuildHasher> crate::TryExtendOne<(K, V)> for HashMap<K, V, S> {
+    type Error = crate::ItemCollision<(K, V)>;
+
+    fn try_extend_one(&mut self, item: (K, V)) -> Result<(), Self::Error> {
+        let hash = self.hasher().hash_one(&item.0);
+        match self.raw_entry_mut().from_hash(hash, |k| k == &item.0) {
+            RawEntryMut::Occupied(_) => Err(crate::ItemCollision::new(item)),
+            RawEntryMut::Vacant(entry) => {
+                entry.insert_hashed_nocheck(hash, item.0, item.1);
+                Ok(())
+            }
+        }
+    }
+}