+++ src/traits/mod.rs
@@ -2,6 +2,6 @@ mod try_extend;
 mod try_from_iterator;
 mod try_unzip;
 
-pub use try_extend::{TryExtend, TryExtendSafe};
-pub use try_from_iterator::{TryCollectEx, TryFromIterator};
+pub use try_extend::*;
+pub use try_from_iterator::*;
 pub use try_unzip::TryUnzip;