+++ tests/arrayvec.rs
@@ -1,15 +1,16 @@
 mod collection_tests;
-use collection_tests::{try_collect, try_extend, try_extend_safe};
 
 use arrayvec::ArrayVec;
 
 use collect_failable::utils::FixedSizeHintEx;
-use collect_failable::{CapacityMismatch, TryExtend, TryExtendSafe, TryFromIterator};
+use collect_failable::CapacityMismatch;
 
 type TestArray = ArrayVec<u32, 2>;
 
 mod try_from_iter {
     use super::*;
+    use collect_failable::TryFromIterator;
+    use collection_tests::try_collect;
 
     try_collect!(valid, TestArray, 1..=2, Ok(ArrayVec::from_iter(1..=2)));
     try_collect!(bounds, TestArray, 1..=3, Err(CapacityMismatch::bounds(0..=2, (3, Some(3)))));
@@ -18,6 +19,8 @@ mod try_from_iter {
 
 #[test]
 fn try_from_iter_reconstruct() {
+    use collect_failable::TryFromIterator;
+
     let err = TestArray::try_from_iter(1..=3).expect_err("Should be err");
 
     let reconstructed: Vec<u32> = err.into_iter().collect();
@@ -29,6 +32,8 @@ fn try_from_iter_reconstruct() {
 
 mod try_extend_safe {
     use super::*;
+    use collect_failable::TryExtendSafe;
+    use collection_tests::try_extend_safe;
 
     try_extend_safe!(valid, TestArray::from_iter(3..=3), 3..=3, Ok(TestArray::from_iter([3, 3])));
     try_extend_safe!(bound_fail, TestArray::from_iter(3..=3), 1..=3, Err(CapacityMismatch::bounds(0..=1, (1..=3).size_hint())));
@@ -37,8 +42,21 @@ mod try_extend_safe {
 
 mod try_extend {
     use super::*;
+    use collect_failable::TryExtend;
+    use collection_tests::try_extend;
 
-    try_extend!(valid, ArrayVec::<u32, 2>::from_iter(3..=3), 3..=3, Ok(ArrayVec::from_iter([3, 3])));
-    try_extend!(bounds_fail, ArrayVec::<u32, 2>::from_iter(3..=3), 1..=3, Err(CapacityMismatch::bounds(0..=1, (3, Some(3)))));
-    try_extend!(overflow, ArrayVec::<u32, 2>::from_iter(3..=3), (1..=2).hide_size(), Err(CapacityMismatch::overflow(0..=1)));
+    try_extend!(valid, TestArray::from_iter(3..=3), 3..=3, Ok(ArrayVec::from_iter([3, 3])));
+    try_extend!(bounds_fail, TestArray::from_iter(3..=3), 1..=3, Err(CapacityMismatch::bounds(0..=1, (3, Some(3)))));
+    try_extend!(overflow, TestArray::from_iter(3..=3), (1..=2).hide_size(), Err(CapacityMismatch::overflow(0..=1)));
+}
+
+mod try_extend_one {
+    use arrayvec::CapacityError;
+    use collect_failable::TryExtendOne;
+    use collection_tests::try_extend_one;
+
+    use super::*;
+
+    try_extend_one!(valid, TestArray::new(), 1, Ok(TestArray::from_iter(1..=1)));
+    try_extend_one!(collision, TestArray::from([1, 2]), 2, Err(CapacityError::new(2)));
 }