+++ tests/sets.rs
@@ -1,9 +1,12 @@
+mod collection_tests;
+use collection_tests::{try_extend, try_extend_one, try_extend_safe};
+
 use std::collections::{BTreeSet, HashSet};
 
 use hashbrown::HashSet as HashBrownSet;
 
 use collect_failable::utils::Identifiable;
-use collect_failable::{TryExtend, TryExtendSafe, TryFromIterator};
+use collect_failable::{ItemCollision, TryExtend, TryExtendOne, TryExtendSafe, TryFromIterator};
 
 macro_rules! try_from_iter_and_extend_iter {
     ($module:ident, $set_type:ty) => {
@@ -80,14 +83,12 @@ macro_rules! try_from_iter_and_extend_iter {
                 assert_eq!(remaining, vec![5, 3], "remaining should have 2 items");
             }
 
-            #[test]
-            fn try_extend_safe_no_collision() {
-                let mut set = <$set_type>::new();
-
-                set.try_extend_safe(UNIQUE_VALUES).expect("should be ok");
-
-                assert_eq!(set, <$set_type>::from(UNIQUE_VALUES), "should match data");
-            }
+            try_extend_safe!(
+                try_extend_safe_no_collision,
+                <$set_type>::new(),
+                UNIQUE_VALUES,
+                Ok(<$set_type>::from(UNIQUE_VALUES))
+            );
 
             #[test]
             fn try_extend_collision_with_set() {
@@ -133,14 +134,7 @@ macro_rules! try_from_iter_and_extend_iter {
                 assert_eq!(remaining, vec![5, 3], "remaining should have 2 items");
             }
 
-            #[test]
-            fn try_extend_no_collision() {
-                let mut set = <$set_type>::new();
-
-                set.try_extend(UNIQUE_VALUES).expect("should be ok");
-
-                assert_eq!(set, <$set_type>::from(UNIQUE_VALUES), "should match data");
-            }
+            try_extend!(try_extend_no_collision, <$set_type>::new(), UNIQUE_VALUES, Ok(<$set_type>::from(UNIQUE_VALUES)));
 
             #[test]
             fn try_extend_preserves_original_value() {
@@ -159,6 +153,13 @@ macro_rules! try_from_iter_and_extend_iter {
                 // Verify the set still contains the original value (id: 1)
                 let stored = set.iter().next().unwrap();
                 assert_eq!(stored.id, 1, "Set should contain the original value");
+            }
+
+            mod try_extend_one {
+                use super::*;
+
+                try_extend_one!(valid, <$set_type>::new(), 1, Ok(<$set_type>::from([1])));
+                try_extend_one!(collision, <$set_type>::from([1, 2, 3]), 2, Err(ItemCollision::new(2)));
             }
         }
     };